Merge-DropDown

+++ app/src/main/java/com/creative/weather_app/model/weatherResponse.kt
@@ -16,8 +16,8 @@ data class weatherResponse(
         val timezone: Int
     ) {
         data class Coord(
-            val lat: Int,
-            val lon: Int
+            val lat: Double,
+            val lon: Double
         )
     }
